Merge gwen2_aq_3 branch back onto trunk.
Merged -rHEAD:203 of trunk to undo the libchipcard2 changes of trunk.
Then merged -r203:343 of aq_3 branch to merge the libchipcard3 development.
Then merged -r343:HEAD of gwen2_aq_3 branch so that we have all 
libchipcard3+gwen2 development on trunk now.
The aq_3 branch continues to be the branch for gwen3.

git-svn-id: https://devel.aqbanking.de/svn/libchipcard/trunk@376 bfc04c16-8f10-0410-8200-ddfbc4d55d47